kleine fout in controller type.php bij opslaan van nieuw type, opgelost met if-statement

+++ admin/controllers/location.php
@@ -10,23 +10,25 @@ class JexBookingControllerLocation extends JControllerForm
 	
 		$item_id = $data['id'];
 	
-		$db = JFactory::getDbo();
-		$query = $db->getQuery(true);
-		$query->delete('#__jexbooking_xref_attributes')->where('location_id = '.$item_id);
-		$db->setQuery($query);
-		$db->query();
-		$db->clear();
-	
-	
-		foreach($data['attribute_id'] as $attribute_id){
+		if($item_id){		
 			$db = JFactory::getDbo();
-			$db->clear();
 			$query = $db->getQuery(true);
-			$query->insert('#__jexbooking_xref_attributes');
-			$query->columns('location_id, attribute_id');
-			$query->values("$item_id, $attribute_id");
+			$query->delete('#__jexbooking_xref_attributes')->where('location_id = '.$item_id);
 			$db->setQuery($query);
 			$db->query();
+			$db->clear();
+		
+		
+			foreach($data['attribute_id'] as $attribute_id){
+				$db = JFactory::getDbo();
+				$db->clear();
+				$query = $db->getQuery(true);
+				$query->insert('#__jexbooking_xref_attributes');
+				$query->columns('location_id, attribute_id');
+				$query->values("$item_id, $attribute_id");
+				$db->setQuery($query);
+				$db->query();
+			}
 		}
 	
 		parent::save();
